move tests back into unit tests
